moved app state and ui messaging to separate modules
