+++ src-tauri/src/gui/mod.rs
@@ -19,171 +19,12 @@ use std::{
     thread,
 };
 use tauri::{AppHandle, Manager}; // dep: fork = "0.1"
+mod app_state;
+mod ui;
 
-// Types and structs
-#[derive(Default, Serialize, Deserialize)]
-pub struct AppState {
-    wizard_data: Mutex<WizardData>,
-    settings: Mutex<Settings>,
-    is_installing: Mutex<bool>,
-}
 
-#[tauri::command]
-fn is_installing(app_handle: AppHandle) -> bool {
-    let app_state = app_handle.state::<AppState>();
-    app_state
-        .is_installing
-        .lock()
-        .map(|guard| *guard)
-        .unwrap_or_else(|_| {
-            error!("Failed to acquire is_installing lock, assuming not installing");
-            false
-        })
-}
-
-#[derive(Default, Serialize, Deserialize)]
-struct WizardData {
-    // Add fields relevant to your installation process
-    step_completed: Vec<bool>,
-}
-
-// Event handling
-fn send_message(app_handle: &AppHandle, message: String, message_type: String) {
-    emit_to_fe(
-        app_handle,
-        "user-message",
-        json!({ "type": message_type, "message": message }),
-    );
-}
-
-fn send_tools_message(app_handle: &AppHandle, tool: String, action: String) {
-    emit_to_fe(
-        app_handle,
-        "tools-message",
-        json!({ "tool": tool, "action": action }),
-    );
-}
-
-fn send_install_progress_message(app_handle: &AppHandle, version: String, state: String) {
-    emit_to_fe(
-        app_handle,
-        "install-progress-message",
-        json!({ "version": version, "state": state }),
-    );
-}
-
-fn send_simple_setup_message(app_handle: &AppHandle, message_code: i32, message: String) {
-    emit_to_fe(
-        app_handle,
-        "simple-setup-message",
-        json!({ "code": message_code, "message": message }),
-    );
-}
-
-fn emit_to_fe(app_handle: &AppHandle, event_name: &str, json_data: Value) {
-    debug!("emit_to_fe: {} {:?}", event_name, json_data);
-    let _ = app_handle.emit(event_name, json_data);
-}
-
-#[derive(Clone)]
-struct ProgressBar {
-    app_handle: AppHandle,
-}
-
-impl ProgressBar {
-    fn new(app_handle: AppHandle, message: &str) -> Self {
-        let progress = Self { app_handle };
-        progress.create(message);
-        progress
-    }
-
-    fn create(&self, message: &str) {
-        emit_to_fe(
-            &self.app_handle,
-            "progress-message",
-            json!({
-                "message": message,
-                "status": "info",
-                "percentage": 0,
-                "display": true,
-            }),
-        );
-    }
-
-    fn update(&self, percentage: u64, message: Option<&str>) {
-        emit_to_fe(
-            &self.app_handle,
-            "progress-message",
-            json!({
-                "percentage": percentage,
-                "message": message.unwrap_or_default(),
-                "status": "info",
-                "display": true,
-            }),
-        );
-    }
-
-    fn finish(&self) {
-        info!("finish_progress_bar called");
-        emit_to_fe(
-            &self.app_handle,
-            "progress-message",
-            json!({
-                "message": "",
-                "percentage": 100,
-                "display": false,
-            }),
-        );
-    }
-}
-
-// Settings management
-fn get_locked_settings(app_handle: &AppHandle) -> Result<Settings, String> {
-    let app_state = app_handle.state::<AppState>();
-    app_state
-        .settings
-        .lock()
-        .map(|guard| (*guard).clone())
-        .map_err(|_| {
-            "Failed to obtain lock on AppState. Please retry the last action later.".to_string()
-        })
-}
-
-fn get_settings_non_blocking(app_handle: &AppHandle) -> Result<Settings, String> {
-    let app_state = app_handle.state::<AppState>();
-
-    // First try with a non-blocking try_lock
-    if let Ok(guard) = app_state.settings.try_lock() {
-        let settings_copy = (*guard).clone();
-        return Ok(settings_copy);
-    }
-
-    // If we couldn't get the lock immediately, wait a little and retry
-    for _ in 0..5 {
-        // Small sleep to avoid busy waiting
-        std::thread::sleep(std::time::Duration::from_millis(10));
-
-        if let Ok(guard) = app_state.settings.try_lock() {
-            let settings_copy = (*guard).clone();
-            return Ok(settings_copy);
-        }
-    }
-
-    Err("Settings are currently locked. Try again later.".to_string())
-}
-
-fn update_settings<F>(app_handle: &AppHandle, updater: F) -> Result<(), String>
-where
-    F: FnOnce(&mut Settings),
-{
-    let app_state = app_handle.state::<AppState>();
-    let mut settings = app_state.settings.lock().map_err(|_| {
-        "Failed to obtain lock on AppState. Please retry the last action later.".to_string()
-    })?;
-    updater(&mut settings);
-    debug!("Settings after update: {:?}", settings);
-    Ok(())
-}
+use app_state::{AppState, get_locked_settings, get_settings_non_blocking, update_settings};
+use ui::{send_message,send_install_progress_message, send_simple_setup_message,send_tools_message, emit_to_fe, ProgressBar};
 
 // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
 #[tauri::command]
